requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cssselect2==0.8.0
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
fonttools==4.56.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
Markdown==3.7
mercadopago==2.2.3
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==10.4.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.34.0
weasyprint==64.1
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
